libmbedtls: mbedtls_mpi_exp_mod(): optimize mempool usage
With W allocated in the mempool instead of the stack it is more important
to utilize the mempool in a stack like way.

With this patch allocation and initialization of W is moved to a point
where all following mempool allocations are free before the function
returns.

This reduces maximum memory consumption of mempool in regression case
8101 for an AArch64 TA in from 17280 to 7640 bytes. Figures for an
AArch32 TA are 12040 to 5288 bytes.

Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
[jf: rebased onto mbedtls-2.27.0]
Signed-off-by: Jerome Forissier <jerome@forissier.org>
[jd: rebased onto mbedtls-2.28.2]
[jd: rebased onto mbedtls-3.5.2]
Tracked-On: OAM-115470
Signed-off-by: Jingdong Lu <jingdong.lu@intel.com>