Adjust the auth requirement actions
There are legacy keys which have no authentication-related tags,
so we assume that absence is equivalent to presence of
KM_TAG_NO_AUTH_REQUIRED. So, if we found KM_TAG_USER_AUTH_TYPE
or if we find KM_TAG_USER_SECURE_ID then authentication is
required. If we find neither, then we assume authentication is not
required.

Tracked-On: OAM-123584
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>