QNX: Add vdev shm event source type
Since QNX vdev shm doesn't support msi interrupts, it
supports only one type of interrupt which is passed
to APIC through IO-APIC. To recognize what a interrupt
is used for, need to define a event source type. When
a client trigger an interrupt to the optee, it should
put the event source type to the first four bytes of
the shared memory. optee will fetch the event source
type firstly before handle the interrupt.

Tracked-On: OAM-123671
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>